Tried to get xframe working, failed
To create an axis (the name of the columns and rows, which are the dates and ETF names, respectively), you need to have an initializer array as an input, and I'm not typing 700 dates in manually so this seems like a dead end. Use g++ -I {path_to_project}\portfolio-diversifier\lib\xframe\include\xframe\xtensor\ storedata.cpp -o xframe_test to compile.